refactor(goals): extract goalMath.ts shared utilities (step 9)
Deduplicate getTotalForMonth, getRetirementMonth, monthsBetween,
calcMonthlySaving, and getGwTarget into src/pages/goal/utils/goalMath.ts.
GoalDetail and SavingsPlan now import from shared module.
